add ( v2_simulator_screen ) : ad scaffold
