+++ lib/presentation/screens/catalog/catalog_screen.dart
@@ -46,6 +46,12 @@ class CatalogScreen extends HookConsumerWidget {
           children: [
             ElevatedButton(
               onPressed: () {
+                Navigator.pushNamed(context, '/v2/simulator');
+              },
+              child: const Text('v2 simulador'),
+            ),
+            ElevatedButton(
+              onPressed: () {
                 addVoucherModal(context);
               },
               child: const Text('add voucher'),
