+++ lib/presentation/screens/new_simulator/v2_simulator_screen.dart
@@ -1,0 +1,67 @@
+import 'package:finniu/presentation/providers/settings_provider.dart';
+import 'package:finniu/presentation/screens/business_investments/widgets/app_bar_business.dart';
+import 'package:finniu/presentation/screens/home_v2/widgets/navigation_bar.dart';
+import 'package:finniu/presentation/screens/new_simulator/widgets/icon_found.dart';
+import 'package:finniu/presentation/screens/new_simulator/widgets/investment_amount_card.dart';
+import 'package:finniu/presentation/screens/new_simulator/widgets/investment_ends.dart';
+import 'package:finniu/presentation/screens/new_simulator/widgets/selected_back.dart';
+import 'package:finniu/presentation/screens/new_simulator/widgets/term_profitability_row.dart';
+import 'package:finniu/presentation/screens/new_simulator/widgets/title_simulator.dart';
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+class V2SimulatorScreen extends HookConsumerWidget {
+  const V2SimulatorScreen({super.key});
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final isDarkMode = ref.watch(settingsNotifierProvider).isDarkMode;
+    const int columnColorDark = 0xff0E0E0E;
+    const int columnColorLight = 0xffF8F8F8;
+
+    return Scaffold(
+      backgroundColor: isDarkMode
+          ? const Color(columnColorDark)
+          : const Color(columnColorLight),
+      appBar: const AppBarBusinessScreen(),
+      bottomNavigationBar: const NavigationBarHome(),
+      body: const SingleChildScrollView(
+        child: BodyScaffold(),
+      ),
+    );
+  }
+}
+
+class BodyScaffold extends ConsumerWidget {
+  const BodyScaffold({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final isDarkMode = ref.watch(settingsNotifierProvider).isDarkMode;
+    const int columnColorDark = 0xff0E0E0E;
+    const int columnColorLight = 0xffF8F8F8;
+    return Container(
+      color: isDarkMode
+          ? const Color(columnColorDark)
+          : const Color(columnColorLight),
+      width: MediaQuery.of(context).size.width,
+      height: MediaQuery.of(context).size.height,
+      child: const Padding(
+        padding: EdgeInsets.all(15.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            TitleModal(),
+            IconFund(),
+            InvestmentAmountCardsRow(),
+            TermProfitabilityRow(),
+            SelectedBank(),
+            InvestmentEnds(),
+          ],
+        ),
+      ),
+    );
+  }
+}
