+++ lib/presentation/screens/new_simulator/widgets/selected_back.dart
@@ -1,0 +1,59 @@
+import 'package:finniu/presentation/providers/settings_provider.dart';
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+class SelectedBank extends ConsumerWidget {
+  const SelectedBank({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final isDarkMode = ref.watch(settingsNotifierProvider).isDarkMode;
+    const int backgroundDark = 0xff2A2929;
+    const int backgroundLight = 0xffF1FCFF;
+    return Container(
+      height: 66,
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(10),
+        color: isDarkMode
+            ? const Color(backgroundDark)
+            : const Color(backgroundLight),
+      ),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.start,
+        children: [
+          const SizedBox(width: 13),
+          Image.asset(
+            "assets/images/bankSelectImage.png",
+            width: 45,
+            height: 45,
+          ),
+          const SizedBox(width: 10),
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Text(
+                "Banco donde se transfiere",
+                style: TextStyle(
+                  color: isDarkMode ? Colors.white : Colors.black,
+                  fontSize: 14,
+                ),
+              ),
+              Text(
+                "BCP *************321",
+                style: TextStyle(
+                  color: isDarkMode ? Colors.white : Colors.black,
+                  fontSize: 16,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
