+++ lib/routes/routes.dart
@@ -6,6 +6,7 @@ import 'package:finniu/presentation/screens/home_v2/home_screen.dart';
 import 'package:finniu/presentation/screens/investment_process.dart/step_1_screen.dart';
 import 'package:finniu/presentation/screens/investment_process.dart/step_2_screen.dart';
 import 'package:finniu/presentation/screens/investment_process_blue_gold/investment_blue_gold_screen.dart';
+import 'package:finniu/presentation/screens/new_simulator/v2_simulator_screen.dart';
 import 'package:finniu/presentation/screens/reinvest_process/reinvestment_experience_eval.dart';
 import 'package:finniu/presentation/screens/reinvest_process/reinvestment_step_2.dart';
 import 'package:finniu/presentation/screens/signup/activate_account.dart';
@@ -113,5 +114,6 @@ Map<String, WidgetBuilder> getApplicationRoutes() {
         const BlueGoldInvestmentsScreen(),
     '/v2/investment_blue_gold': (BuildContext context) =>
         const InvestmentBlueGoldScreen(),
+    '/v2/simulator': (BuildContext context) => const V2SimulatorScreen(),
   };
 }
